Fix server app and client app shutdown
Send "exit" command from SplitWise client application to server
in order to ensure connection is closed on server side.

Add check if user input is ready to be read before invoking
reader.readline() on user's input reader. This way ensures
that when SplitWise application is stopped on the server
side the client applications are stopped rather than blocked
on waiting user's IO operation to complete.